fix: refactor rke2 package, will become too big soon for 1 file.
