+++ cmd/rke2.go
@@ -10,11 +10,11 @@ import (
 	"os"
 
 	common "github.com/michielvha/edgectl/pkg/common"
-	lb "github.com/michielvha/edgectl/pkg/lb"
 	"github.com/michielvha/edgectl/pkg/logger"
-	server "github.com/michielvha/edgectl/pkg/rke2/server" // Import the new package
+	server "github.com/michielvha/edgectl/pkg/rke2/server"
 	vault "github.com/michielvha/edgectl/pkg/vault"
 	"github.com/spf13/cobra"
+	lbcmd "github.com/michielvha/edgectl/cmd/rke2/lb"
 )
 
 // rke2Cmd represents the "rke2" command
@@ -114,7 +114,7 @@ var uninstallCmd = &cobra.Command{
 }
 
 // Configure kubeconfig
-var SetKubeConfigCmd = &cobra.Command{
+var setKubeConfigCmd = &cobra.Command{
 	Use:   "config",
 	Short: "Fetch kubeconfig from Vault and store it on the host",
 	Run: func(cmd *cobra.Command, args []string) {
@@ -140,88 +140,15 @@ var SetKubeConfigCmd = &cobra.Command{
 	},
 }
 
-// LB commands
-var lbCmd = &cobra.Command{
-	Use:   "lb",
-	Short: "Manage RKE2 load balancer",
-	Long: `The "lb" command allows you to set up and manage HAProxy load balancers for RKE2.
-	
-Examples:
-  edgectl rke2 lb create --cluster-id my-cluster --vip 192.168.10.100  # Create a new load balancer
-  edgectl rke2 lb status --cluster-id my-cluster                       # Check load balancer status
-`,
-}
-
-// Create LB command
-var lbCreateCmd = &cobra.Command{
-	Use:   "create",
-	Short: "Create a load balancer for RKE2",
+// New command to just configure the bash environment
+var configBashCmd = &cobra.Command{
+	Use:   "config bash",
+	Short: "Configure the bash environment for RKE2",
+	Long:  `Configures the bash environment to use RKE2 binaries and kubeconfig.`,
 	Run: func(cmd *cobra.Command, args []string) {
-		logger.Debug("lb create command executed")
-
-		// Check if user is root
-		if os.Geteuid() != 0 {
-			fmt.Println("❌ This command must be run as root. Try using `sudo`.")
-			os.Exit(1)
-		}
-
-		// Extract values
-		clusterID, _ := cmd.Flags().GetString("cluster-id")
-		vip, _ := cmd.Flags().GetString("vip")
-
-		// Create load balancer
-		err := lb.CreateLoadBalancer(clusterID, vip)
-		if err != nil {
-			fmt.Printf("❌ Failed to create load balancer: %v\n", err)
-			os.Exit(1)
-		}
-
-		fmt.Println("✅ RKE2 load balancer created successfully")
-	},
-}
-
-// LB Status command
-var lbStatusCmd = &cobra.Command{
-	Use:   "status",
-	Short: "Show status of RKE2 load balancer",
-	Run: func(cmd *cobra.Command, args []string) {
-		logger.Debug("lb status command executed")
-
-		clusterID, _ := cmd.Flags().GetString("cluster-id")
-		if clusterID == "" {
-			fmt.Println("❌ Cluster ID is required.")
-			_ = cmd.Help()
-			os.Exit(1)
-		}
-
-		// Connect to Vault
-		client, err := vault.NewClient()
-		if err != nil {
-			fmt.Printf("❌ Failed to create Vault client: %v\n", err)
-			os.Exit(1)
-		}
-
-		// Get LB info
-		lbNodes, vip, err := client.RetrieveLBInfo(clusterID)
-		if err != nil {
-			fmt.Printf("❌ Failed to retrieve load balancer info: %v\n", err)
-			os.Exit(1)
-		}
-
-		fmt.Printf("ℹ️ RKE2 Load balancer VIP: %s\n", vip)
-		fmt.Println("ℹ️ Load balancer nodes:")
-
-		for _, node := range lbNodes {
-			hostname := node["hostname"].(string)
-			isMain := node["is_main"].(bool)
-
-			role := "BACKUP"
-			if isMain {
-				role = "MASTER"
-			}
-
-			fmt.Printf("  - %s (%s)\n", hostname, role)
-		}
+		fmt.Println("🔧 Configuring bash environment for RKE2...")
+		common.RunBashFunction("rke2.sh", "configure_rke2_bash")
+		fmt.Println("✅ Bash environment configured for RKE2")
 	},
 }
 
@@ -238,27 +165,18 @@ func init() {
 	installAgentCmd.Flags().String("lb-hostname", "", "The hostname of the load balancer to use if VIP is not found")
 	_ = installAgentCmd.MarkFlagRequired("cluster-id")
 
-	SetKubeConfigCmd.Flags().String("cluster-id", "", "The ID of the cluster to fetch the kubeconfig for")
-	SetKubeConfigCmd.Flags().String("output", "/etc/rancher/rke2/rke2.yaml", "Destination path to store the kubeconfig")
-	_ = SetKubeConfigCmd.MarkFlagRequired("cluster-id")
-
-	// LB command flags
-	lbCreateCmd.Flags().String("cluster-id", "", "The ID of the cluster to create a load balancer for")
-	lbCreateCmd.Flags().String("vip", "", "Virtual IP address for the load balancer")
-	_ = lbCreateCmd.MarkFlagRequired("cluster-id")
-
-	lbStatusCmd.Flags().String("cluster-id", "", "The ID of the cluster to check load balancer status for")
-	_ = lbStatusCmd.MarkFlagRequired("cluster-id")
-
-	// Add LB commands
-	lbCmd.AddCommand(lbCreateCmd)
-	lbCmd.AddCommand(lbStatusCmd)
-	rke2Cmd.AddCommand(lbCmd)
+	setKubeConfigCmd.Flags().String("cluster-id", "", "The ID of the cluster to fetch the kubeconfig for")
+	setKubeConfigCmd.Flags().String("output", "/etc/rancher/rke2/rke2.yaml", "Destination path to store the kubeconfig")
+	_ = setKubeConfigCmd.MarkFlagRequired("cluster-id")
 
 	// Attach subcommands under rke2
 	rke2Cmd.AddCommand(installServerCmd)
 	rke2Cmd.AddCommand(installAgentCmd)
 	rke2Cmd.AddCommand(statusCmd)
 	rke2Cmd.AddCommand(uninstallCmd)
-	rke2Cmd.AddCommand(SetKubeConfigCmd)
+	rke2Cmd.AddCommand(setKubeConfigCmd)
+	rke2Cmd.AddCommand(configBashCmd)
+
+	// Add loadbalancer command from the new package
+	rke2Cmd.AddCommand(lbcmd.Cmd)
 }
